Separate `testExample` from `DownloaderTests`
